Face_Detection_System/dockerfiles/requirements.txt
pika
opencv-python
psycopg2
python-dotenv
mediapipe